--- launcher-package/integration-test/src/test/scala/PowerAssertions.scala
@@ -1,7 +1,0 @@
-package example.test
-
-import com.eed3si9n.expecty.Expecty
-
-trait PowerAssertions {
-  lazy val assert: Expecty = new Expecty()
-}
