--- launcher-package/project/plugins.sbt
@@ -1,2 +1,0 @@
-addSbtPlugin("com.github.sbt" % "sbt-native-packager" % "1.10.4")
-addSbtPlugin("com.github.sbt" % "sbt-pgp" % "2.3.1")
