+++ project/PackageSignerPlugin.scala
@@ -1,5 +1,5 @@
-import sbt._
-import Keys._
+import sbt.*
+import Keys.*
 import sbt.internal.librarymanagement.IvyActions
 import com.jsuereth.sbtpgp.SbtPgp
 import com.typesafe.sbt.packager.universal.{ UniversalPlugin, UniversalDeployPlugin }
@@ -11,20 +11,20 @@ object PackageSignerPlugin extends sbt.AutoPlugin {
   override def trigger = allRequirements
   override def requires = SbtPgp && UniversalDeployPlugin && DebianDeployPlugin && RpmDeployPlugin
 
-  import com.jsuereth.sbtpgp.PgpKeys._
-  import UniversalPlugin.autoImport._
-  import DebianPlugin.autoImport._
-  import RpmPlugin.autoImport._
+  import com.jsuereth.sbtpgp.PgpKeys.*
+  import UniversalPlugin.autoImport.*
+  import DebianPlugin.autoImport.*
+  import RpmPlugin.autoImport.*
 
-  override def projectSettings: Seq[Setting[_]] =
+  override def projectSettings: Seq[Setting[?]] =
     inConfig(Universal)(packageSignerSettings) ++
-    inConfig(Debian)(packageSignerSettings) ++
-    inConfig(Rpm)(packageSignerSettings)
+      inConfig(Debian)(packageSignerSettings) ++
+      inConfig(Rpm)(packageSignerSettings)
 
   def subExtension(art: Artifact, ext: String): Artifact =
     art.withExtension(ext)
 
-  def packageSignerSettings: Seq[Setting[_]] = Seq(
+  def packageSignerSettings: Seq[Setting[?]] = Seq(
     signedArtifacts := {
       val artifacts = packagedArtifacts.value
       val r = pgpSigner.value
@@ -32,33 +32,38 @@ object PackageSignerPlugin extends sbt.AutoPlugin {
       val s = streams.value
       if (!skipZ) {
         artifacts flatMap { case (art, f) =>
-          Seq(art -> f,
-              subExtension(art, art.extension + gpgExtension) ->
-              r.sign(f, file(f.getAbsolutePath + gpgExtension), s))
+          Seq(
+            art -> f,
+            subExtension(art, art.extension + gpgExtension) ->
+              r.sign(f, file(f.getAbsolutePath + gpgExtension), s)
+          )
         }
-      }
-      else artifacts
+      } else artifacts
     },
     publishSignedConfiguration := Classpaths.publishConfig(
       publishMavenStyle = publishMavenStyle.value,
-      deliverIvyPattern = (Compile / packageBin / artifactPath).value.getParent + "/[artifact]-[revision](-[classifier]).[ext]",
+      deliverIvyPattern =
+        (Compile / packageBin / artifactPath).value.getParent + "/[artifact]-[revision](-[classifier]).[ext]",
       status = if (isSnapshot.value) "integration" else "release",
       configurations = Vector.empty,
       artifacts = signedArtifacts.value.toVector,
       checksums = (publish / checksums).value.toVector,
       resolverName = Classpaths.getPublishTo(publishTo.value).name,
       logging = ivyLoggingLevel.value,
-      overwrite = isSnapshot.value),
+      overwrite = isSnapshot.value
+    ),
     publishLocalSignedConfiguration := Classpaths.publishConfig(
       publishMavenStyle = publishMavenStyle.value,
-      deliverIvyPattern = (Compile / packageBin / artifactPath).value.getParent + "/[artifact]-[revision](-[classifier]).[ext]",
+      deliverIvyPattern =
+        (Compile / packageBin / artifactPath).value.getParent + "/[artifact]-[revision](-[classifier]).[ext]",
       status = if (isSnapshot.value) "integration" else "release",
       configurations = Vector.empty,
       artifacts = signedArtifacts.value.toVector,
       checksums = (publish / checksums).value.toVector,
       resolverName = "local",
       logging = ivyLoggingLevel.value,
-      overwrite = isSnapshot.value),
+      overwrite = isSnapshot.value
+    ),
     publishSigned := Def.taskDyn {
       val config = publishSignedConfiguration.value
       val s = streams.value
@@ -76,4 +81,3 @@ object PackageSignerPlugin extends sbt.AutoPlugin {
   )
 
 }
-
