+++ project/plugins.sbt
@@ -11,5 +11,6 @@ addSbtPlugin("com.swoval" % "sbt-java-format" % "0.3.1")
 addSbtPlugin("org.scalameta" % "sbt-native-image" % "0.3.4")
 addDependencyTreePlugin
 addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.14.3")
+addSbtPlugin("com.github.sbt" % "sbt-native-packager" % "1.10.4")
 
 // libraryDependencies += "org.scala-sbt" %% "scripted-plugin" % sbtVersion.value
